backend/requirements.txt
# Actian VectorAI DB client (vendored wheel)
./vendor/actian_vectorai-0.1.0b2-py3-none-any.whl

# Web framework
fastapi>=0.115,<0.120
uvicorn[standard]>=0.32,<0.35
python-multipart>=0.0.9

# Embedding models
# NOTE: transformers is pinned to the 4.x series — FlagEmbedding 1.3.x uses
# the `is_torch_fx_available` symbol which was removed in transformers 5.0.
FlagEmbedding>=1.3.0,<2.0
sentence-transformers>=3.0.0,<6.0
transformers>=4.44.0,<5.0
torch>=2.3.0,<3.0

# Data + images
pandas>=2.2.0
numpy>=1.26.0
pillow>=10.4.0

# HTTP + utilities
requests>=2.32.0
python-dotenv>=1.0.1

# Tests
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.27.0

# Stretch (Task 2.8 BM25 path — safe to install now)
rank-bm25>=0.2.2
scikit-learn>=1.5.0
